requirements.txt
Django==4.2.6
django-compression-middleware==0.4.1
django-cors-headers==3.10.1
django-environ==0.9.0
django-filter==21.1
django-jsonfield==1.4.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
drf-nested-routers==0.93.4
requests==2.31.0
urllib3==1.26.18
cffi==1.16.0
psycopg2-binary==2.9.9
pycparser==2.21
Pillow==10.1.0
uvloop==0.19.0
colorlog==6.7.0 
uWSGI==2.0.22
whitenoise==6.6.0